# docs/swagger.yaml
basePath: /api
definitions:
...
  handlers.JournalAssetDTO:
    properties:
      filename:
        type: string
      path:
        type: string
      size_bytes:
        type: integer
    type: object
  handlers.JournalDTO:
    properties:
      created_at:
        type: string
      day:
        type: integer
      hash:
        type: string
      last_checked_at:
        type: string
      month:
        type: integer
      size_bytes:
        type: integer
      tags:
        items:
          type: string
        type: array
      updated_at:
        type: string
      year:
        type: integer
    type: object
  handlers.JournalDayDTO:
    properties:
      day:
        type: integer
      month:
        type: integer
      raw:
        type: string
      year:
        type: integer
    type: object
...
  handlers.createJournalEntryRequest:
    properties:
      body:
        type: string
      description:
        type: string
      tags:
        items:
          type: string
        type: array
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Local network API for music management
  title: Music Server API
  version: "1.0"
paths:
...
  /journals:
    get:
      parameters:
      - description: Filter by year
        in: query
        name: year
        type: integer
      - description: Filter by month (1-12)
        in: query
        name: month
        type: integer
      - description: Filter by day (1-31)
        in: query
        name: day
        type: integer
      - description: Filter by tag
        in: query
        name: tag
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handlers.JournalDTO'
            type: array
      summary: List journals
      tags:
      - journals
    post:
      consumes:
      - application/json
      operationId: createJournalEntry
      parameters:
      - description: Journal entry payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.createJournalEntryRequest'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
      summary: Append a journal entry for today
      tags:
      - journals
  /journals/{year}/{month}:
    get:
      operationId: listJournalsByMonth
      parameters:
      - description: Force refresh from disk
        in: query
        name: refresh
        type: boolean
      - description: Year
        in: path
        name: year
        required: true
        type: integer
      - description: Month
        in: path
        name: month
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handlers.JournalDTO'
            type: array
      summary: List journals for a month
      tags:
      - journals
  /journals/{year}/{month}/{day}:
    get:
      operationId: getJournalDay
      parameters:
      - description: Year
        in: path
        name: year
        required: true
        type: integer
      - description: Month
        in: path
        name: month
        required: true
        type: integer
      - description: Day
        in: path
        name: day
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.JournalDayDTO'
      summary: Get journal entry for a day
      tags:
      - journals
  /journals/assets:
    get:
      parameters:
      - description: Asset path or filename
        in: query
        name: path
        required: true
        type: string
      produces:
      - application/octet-stream
      responses:
        "200":
          description: OK
      summary: Get journal asset
      tags:
      - journals
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: Asset file
        in: formData
        name: file
        required: true
        type: file
      - description: Override filename
        in: query
        name: filename
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handlers.JournalAssetDTO'
      summary: Upload journal asset
      tags:
      - journals
...
swagger: "2.0"